requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
WTForms==3.0.1
MySQL-connector-python==8.1.0
PyMySQL==1.1.0
passlib==1.7.4
email-validator==2.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
PyDrive2==1.12.0
google-api-python-client==2.88.0
oauth2client==4.1.3
python-dotenv==1.0.1
reportlab==4.0.4
Authlib==1.3.1
cryptography==41.0.7
itsdangerous==2.1.2
requests==2.31.0